pyproject.toml
[tool.poetry]
name = "marp-pptx"
version = "0.1.0"
description = ""
authors = ["Takeo Imai <takeo.bono@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-pptx = "^1.0.2"
python-dotenv = "^1.0.1"
mypy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
